User-id based sharding implemented.
First committed Jul 20, 2006 (508e48603498a5915fdf94316691ac046).

Committers:
   935  megan
   905  timo
   349  pdrapeau
   255  baxter
   193  daria
   172  nathan
    87  travis
    48  thomas

git-svn-id: file:///var/svn/nexopia/trunk@3553 7be605d3-0650-4ba7-bb8c-0af3a60b5258

+++ php-site/include/usernotify.php
@@ -2,72 +2,100 @@
 
 class usernotify
 {
-	var $db;
-	
-	function usernotify(&$db)
+	public $db;
+
+	function __construct(&$db)
 	{
 		$this->db = &$db;
 	}
 	function listNotifications($userid = false, $pagenum = false)
 	{
-		global $userData, $pagenum, $config;
-		
+		global $userData, $config;
+
 		if ($userid === false)
 			$userid = $userData['userid'];
-		
+
 		$queryLimit = "";
 		if ($pagenum === false)
 			$queryLimit .= " LIMIT " . $pagenum*$config['linesPerPage'] . ", " . $config['linesPerPage'];
+
+		$result = $this->db->prepare_query("SELECT usernotifyid, creatorid, createtime, triggertime, subject FROM usernotify WHERE targetid = # ORDER BY triggertime DESC " . $queryLimit, $userid);
+
+		$rows = $result->fetchrowset();
 		
-		$result = $this->db->prepare_query("SELECT usernotifyid, creatorid, creatorname, createtime, triggertime, subject FROM usernotify WHERE targetid = ? ORDER BY triggertime DESC " . $queryLimit, $userid);
+		$uids = array();
+		foreach($rows as $row)
+			$uids[$row['creatorid']] = $row['creatorid'];
 		
-		return $this->db->fetchrowset($result);
+		$usernames = getUserName($uids);
+		
+		foreach($rows as $k => $v)
+			$rows[$k]['creatorname'] = $usernames[$v['creatorid']];
+
+		return $rows;
 	}
-	
-	function newNotify($targetid, $targetname, $triggertime, $subject, $message, $creatorid = false, $creatorname = false)
+
+	function getNotification($notifyid)
+	{
+		$result = $this->db->prepare_query("SELECT * FROM usernotify WHERE usernotifyid = #", $notifyid);
+		
+		$row = $result->fetchrow();
+		
+		$row['creatorname'] = getUserName($row['creatorid']);
+		
+		return $row;
+	}
+
+	function newNotify($targetid, $triggertime, $subject, $message, $creatorid = false)
 	{
 		global $userData;
-		
-		if ($creatorid === false || $creatorname === false)
-		{
+
+		if ($creatorid === false)
 			$creatorid = $userData['userid'];
-			$creatorname = $userData['username'];
-		}
-		
+
 		$this->db->prepare_query(
-			'INSERT INTO usernotify SET creatorid = #, creatorname = ?, createtime = #, targetid = #, triggertime = #, subject = ?, message = ?',
-			                           $creatorid,    $creatorname,     time(),        $targetid,    $triggertime,    $subject,    $message);
+			'INSERT INTO usernotify SET creatorid = #, createtime = #, targetid = #, triggertime = #, subject = ?, message = ?',
+			                           $creatorid,     time(),        $targetid,    $triggertime,    $subject,    $message);
+		return $this->db->affectedrows();
 	}
-	
+
 	function deleteNotify($usernotifyids)
 	{
-		$this->db->prepare_query('DELETE FROM usernotify WHERE usernotifyid IN (#)', $usernotifyids);
+		if ($usernotifyids)
+		{
+			$this->db->prepare_query('DELETE FROM usernotify WHERE usernotifyid IN (#)', $usernotifyids);
+			return $this->db->affectedrows();
+		} else {
+			return 0;
+		}
 	}
-	
+
 	function triggerNotifications()
 	{
 		global $messaging;
-		
+
 		$currenttime = time();
-		
+
 		// fetch notifies with a triggertime in the past
-		$result = $this->db->prepare_query('SELECT creatorid, creatorname, targetid, subject, message FROM usernotify WHERE triggertime <= #', $currenttime);
-		
+		$result = $this->db->prepare_query('SELECT creatorid, targetid, subject, message FROM usernotify WHERE triggertime <= #', $currenttime);
+
 		// send them out as messages
 		$count = 0;
-		while ($triggered = $this->db->fetchrow($result))
+		while ($triggered = $result->fetchrow())
 		{
 			$count++;
 			if ($triggered['creatorid'] == $triggered['targetid'])
 				$triggered['creatorname'] = 'yourself';
-			
+			else
+				$triggered['creatorname'] = getUserName($triggered['creatorid']);
+
 			$triggered['subject'] = "Timed Notification from " . $triggered['creatorname'] . ": " . $triggered['subject'];
 			$messaging->deliverMsg($triggered['targetid'], $triggered['subject'], $triggered['message'], 0, "Nexopia", 0);
 		}
-		
+
 		// now remove them from the database
 		$this->db->prepare_query('DELETE FROM usernotify WHERE triggertime <= #', $currenttime);
-		
+
 		return $count;
 	}
 }
